+++ core/src/main/java/de/bitbrain/braingdx/tmx/TiledMapManager.java
@@ -30,8 +30,6 @@ import de.bitbrain.braingdx.ai.pathfinding.PathFinder;
  */
 public interface TiledMapManager extends Disposable {
 
-   void addListener(TiledMapListener listener);
-
    void load(TiledMap tiledMap, Camera camera, TiledMapType type, TiledMapConfig config) throws TiledMapException;
 
    void load(TiledMap tiledMap, Camera camera, TiledMapType type) throws TiledMapException;