+++ demo/core/src/main/java/de/bitbrain/braingdx/apps/tmxgame/TmxScreen.java
@@ -8,19 +8,18 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.utils.viewport.FillViewport;
-import com.badlogic.gdx.utils.viewport.FitViewport;
-import com.badlogic.gdx.utils.viewport.ScreenViewport;
-import com.badlogic.gdx.utils.viewport.StretchViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
 
 import de.bitbrain.braingdx.GameContext;
 import de.bitbrain.braingdx.ai.pathfinding.Path;
 import de.bitbrain.braingdx.apps.Assets;
-import de.bitbrain.braingdx.apps.Assets.Sounds;
 import de.bitbrain.braingdx.apps.rpg.NPC;
 import de.bitbrain.braingdx.assets.SharedAssetManager;
 import de.bitbrain.braingdx.behavior.movement.Orientation;
 import de.bitbrain.braingdx.behavior.movement.RasteredMovementBehavior;
+import de.bitbrain.braingdx.event.GameEvent;
+import de.bitbrain.braingdx.event.GameEventListener;
+import de.bitbrain.braingdx.event.GameEventManager;
 import de.bitbrain.braingdx.graphics.GraphicsFactory;
 import de.bitbrain.braingdx.graphics.animation.SpriteSheet;
 import de.bitbrain.braingdx.graphics.animation.SpriteSheetAnimation;
@@ -34,24 +33,25 @@ import de.bitbrain.braingdx.graphics.pipeline.layers.RenderPipeIds;
 import de.bitbrain.braingdx.graphics.renderer.AnimationRenderer;
 import de.bitbrain.braingdx.input.OrientationMovementController;
 import de.bitbrain.braingdx.screens.AbstractScreen;
-import de.bitbrain.braingdx.tmx.IndexCalculator;
 import de.bitbrain.braingdx.tmx.TiledMapAPI;
-import de.bitbrain.braingdx.tmx.TiledMapListenerAdapter;
+import de.bitbrain.braingdx.tmx.TiledMapEvents;
 import de.bitbrain.braingdx.tmx.TiledMapManager;
 import de.bitbrain.braingdx.tmx.TiledMapType;
 import de.bitbrain.braingdx.world.GameObject;
 
 public class TmxScreen extends AbstractScreen<TmxTest> {
    
-   private class AStarRenderer extends TiledMapListenerAdapter implements RenderLayer {
+   private class AStarRenderer implements RenderLayer {
       
       private final TiledMapManager tiledMapManager;
       private Path path;
       private Texture texture;
       
-      public AStarRenderer(TiledMapManager manager) {
+      AStarRenderer(TiledMapManager manager, GameEventManager eventManager) {
          this.tiledMapManager = manager;
          texture = GraphicsFactory.createTexture(2, 2, Color.GREEN);
+         eventManager.register(onEnterCellListener, TiledMapEvents.OnEnterCellEvent.class);
+         eventManager.register(onLayerChangeListener, TiledMapEvents.OnLayerChangeEvent.class);
       }
 
       @Override
@@ -74,17 +74,20 @@ public class TmxScreen extends AbstractScreen<TmxTest> {
             batch.end();
          }
       }
-      
-      @Override
-      public void onEnterCell(int xIndex, int yIndex, GameObject object, TiledMapAPI api) {
-         path = tiledMapManager.getPathFinder().findPath(player, 0, 0);
-      }
-      
-      @Override
-      public void onLayerChange(int previousLayer, int newLayer, GameObject object, TiledMapAPI api) {
-         path = tiledMapManager.getPathFinder().findPath(player, 0, 0);
-      }
-      
+
+      private final GameEventListener<TiledMapEvents.OnEnterCellEvent> onEnterCellListener = new GameEventListener<TiledMapEvents.OnEnterCellEvent>() {
+         @Override
+         public void onEvent(TiledMapEvents.OnEnterCellEvent event) {
+            path = tiledMapManager.getPathFinder().findPath(player, 0, 0);
+         }
+      };
+
+      private final GameEventListener<TiledMapEvents.OnLayerChangeEvent> onLayerChangeListener = new GameEventListener<TiledMapEvents.OnLayerChangeEvent>() {
+         @Override
+         public void onEvent(TiledMapEvents.OnLayerChangeEvent event) {
+            path = tiledMapManager.getPathFinder().findPath(player, 0, 0);
+         }
+      };
    }
 
    private GameObject player;
@@ -125,8 +128,7 @@ public class TmxScreen extends AbstractScreen<TmxTest> {
             .rasterSize(tiledMapManager.getAPI().getCellWidth(), tiledMapManager.getAPI().getCellHeight());
       context.getBehaviorManager().apply(behavior, player);
       
-      AStarRenderer renderer = new AStarRenderer(context.getTiledMapManager());
-      context.getTiledMapManager().addListener(renderer);
+      AStarRenderer renderer = new AStarRenderer(context.getTiledMapManager(), context.getEventManager());
       context.getRenderPipeline().putAfter(RenderPipeIds.LIGHTING, "astar", renderer);
    }